pyproject.toml
[tool.poetry]
name = "app-inventor-api"
version = "0.1.0"
description = ""
authors = ["CoolJosh0221 <joshluffy99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
icecream = "^2.1.3"
python-dotenv = "^1.0.0"
gtts = "^2.4.0"
google-api-python-client = "^2.105.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.9.1"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
